+++ examples/minimal/main.tf
@@ -1,5 +1,5 @@
 module "rds_postgres" {
-  # source               = "git::https://github.com/Datatamer/terraform-aws-rds-postgres.git?ref=2.0.0"
+  # source               = "git::https://github.com/Datatamer/terraform-aws-rds-postgres.git?ref=3.0.0"
   source = "../.."
 
   identifier_prefix    = "example-rds-pg-"
@@ -13,10 +13,11 @@ module "rds_postgres" {
   # Network requirement: DB subnet group needs a subnet in at least two Availability Zones
   rds_subnet_ids     = var.subnet_ids
   security_group_ids = module.rds-postgres-sg.security_group_ids
+  tags               = var.tags
 }
 
 module "sg-ports" {
-  # source               = "git::https://github.com/Datatamer/terraform-aws-rds-postgres.git//modules/rds-postgres-ports?ref=2.0.0"
+  # source               = "git::https://github.com/Datatamer/terraform-aws-rds-postgres.git//modules/rds-postgres-ports?ref=3.0.0"
   source = "../../modules/rds-postgres-ports"
 }
 
@@ -29,4 +30,5 @@ module "rds-postgres-sg" {
   sg_name_prefix      = var.name_prefix
   egress_protocol     = "all"
   ingress_protocol    = "tcp"
+  tags                = var.tags
 }