+++ main.tf
@@ -1,13 +1,18 @@
+locals {
+  effective_tags = length(var.tags) > 0 ? var.tags : var.additional_tags
+}
+
 resource "aws_db_parameter_group" "rds_postgres_pg" {
   name        = var.parameter_group_name
   family      = var.parameter_group_family
   description = "TAMR RDS parameter group"
-  tags        = var.additional_tags
+  tags        = local.effective_tags
 }
 
 resource "aws_db_subnet_group" "rds_postgres_subnet_group" {
   name       = var.subnet_group_name
   subnet_ids = var.rds_subnet_ids
+  tags       = local.effective_tags
 }
 
 resource "aws_db_instance" "rds_postgres" {
@@ -19,9 +24,10 @@ resource "aws_db_instance" "rds_postgres" {
   storage_type          = var.storage_type
   storage_encrypted     = true
 
-  engine         = "postgres"
-  engine_version = var.engine_version
-  instance_class = var.instance_class
+  engine                     = "postgres"
+  engine_version             = var.engine_version
+  instance_class             = var.instance_class
+  auto_minor_version_upgrade = var.auto_minor_version_upgrade
 
   username = var.username
   password = var.password
@@ -41,7 +47,7 @@ resource "aws_db_instance" "rds_postgres" {
   apply_immediately = var.apply_immediately
 
   copy_tags_to_snapshot = var.copy_tags_to_snapshot
-  tags                  = var.additional_tags
+  tags                  = local.effective_tags
 
   lifecycle {
     ignore_changes = [password]